Organize data model files under cartography.models (#1096)
* Organize model files under cartography.models

* Update with upstream cleanup changes
