+++ cartography/graph/job.py
@@ -13,10 +13,10 @@ from typing import Union
 import neo4j
 
 from cartography.graph.cleanupbuilder import build_cleanup_queries
-from cartography.graph.model import CartographyNodeSchema
-from cartography.graph.model import CartographyRelSchema
 from cartography.graph.statement import get_job_shortname
 from cartography.graph.statement import GraphStatement
+from cartography.models.core.nodes import CartographyNodeSchema
+from cartography.models.core.relationships import CartographyRelSchema
 
 logger = logging.getLogger(__name__)
