+++ tests/data/graph/querybuilder/sample_models/asset_with_non_kwargs_tgm.py
@@ -1,12 +1,12 @@
 from dataclasses import dataclass
 
-from cartography.graph.model import CartographyNodeSchema
-from cartography.graph.model import CartographyRelProperties
-from cartography.graph.model import CartographyRelSchema
-from cartography.graph.model import LinkDirection
-from cartography.graph.model import make_target_node_matcher
-from cartography.graph.model import PropertyRef
-from cartography.graph.model import TargetNodeMatcher
+from cartography.models.core.common import PropertyRef
+from cartography.models.core.nodes import CartographyNodeSchema
+from cartography.models.core.relationships import CartographyRelProperties
+from cartography.models.core.relationships import CartographyRelSchema
+from cartography.models.core.relationships import LinkDirection
+from cartography.models.core.relationships import make_target_node_matcher
+from cartography.models.core.relationships import TargetNodeMatcher
 from tests.data.graph.querybuilder.sample_models.simple_node import SimpleNodeProperties
